Refactoring xml generation into classes   - Adding two classes one called XMLGeneration that takes care of initializing the generation and an Elements class that generates the element nodes within the document   - Adding tests for top level class of module.
