+++ spec/fixtures/lib/hyacinth/xml_generator/test_dynamic_field_data.json
